Chapter 3: Even cruftier text styling!
I need to clean the code up and try and remove all the odd whitespace
artifacts that are present so that the rendered page looks closer to the
one he has. I've been ignoring it but we might get into more in depth
styling before we right a proper parser so... oh well!

I should also think about splitting the code into multiple files!