Chapter 4: Constructing an HTML Tree
There is a memory leak because I am not using smart pointers for DOM
nodes. I'm not entirely sure what type of pointer I should use, each
child is uniquely owned by it's parent, but each child has a parent
pointer. I don't think it's right to alias a unique_ptr because even
aliasing a Box in Rust is UB, but a shared_ptr seems so silly because
it's just one instance of a shared_ptr and then a single weak_ptr from
child to parent. :(