Chapter 6:
The movement towards stylesheets has undone some features we implemented
in C++ before. Those are going to be reconsidered at a later date. The
fact that we are using Skia instead of Tkinter is also becoming a small
pain for colors. Looking ahead I've built the color table, but haven't
implemented it all quite correctly cause the code is only halfway there.
I think it'll be alright for now.