Handle \r\n split across network reads for chunked transfer
We were in a case where we were asserting that we didn't have a split
\r\n, but we didn't, we just had an \r at the end of the input. This
caused us to crash when we didn't need to.

I'm still all for these assertions, because I'd rather crash prematurely
than ignore an issue that actually happens. This is sort of anti-the-web
culture but I think it's ok for this learning project.